Fixed compiler warnings

+++ src/slic3r/GUI/GCodeViewer.cpp
@@ -3449,7 +3449,7 @@ void GCodeViewer::render_legend(float& legend_height)
 
             std::vector<CustomGCode::Item> custom_gcode_per_print_z = wxGetApp().is_editor() ? wxGetApp().plater()->model().custom_gcode_per_print_z.gcodes : m_custom_gcode_per_print_z;
             std::vector<Color> last_color(m_extruders_count);
-            for (int i = 0; i < m_extruders_count; ++i) {
+            for (size_t i = 0; i < m_extruders_count; ++i) {
                 last_color[i] = m_tool_colors[i];
             }
             int last_extruder_id = 1;
